requirements.txt
browser-use==0.12.6
langchain-openai==1.1.9
python-dotenv==1.2.1
Pillow==12.1.1
